(Qt) Combine all Qt files into one (ui_qt.cpp)
