+++ griffin/griffin_cpp.cpp
@@ -51,11 +51,7 @@ UI
 
 #include "../ui/drivers/ui_qt.cpp"
 
-#include "../ui/drivers/qt/ui_qt_window.cpp"
 #include "../ui/drivers/qt/ui_qt_load_core_window.cpp"
-#include "../ui/drivers/qt/ui_qt_browser_window.cpp"
-#include "../ui/drivers/qt/ui_qt_msg_window.cpp"
-#include "../ui/drivers/qt/ui_qt_application.cpp"
 #include "../ui/drivers/qt/gridview.cpp"
 #if defined(HAVE_CG) || defined(HAVE_GLSL) || defined(HAVE_SLANG) || defined(HAVE_HLSL)
 #include "../ui/drivers/qt/shaderparamsdialog.cpp"
