+++ ui/ui_companion_driver.h
@@ -144,11 +144,6 @@ typedef struct ui_companion_driver
    const char        *ident;
 } ui_companion_driver_t;
 
-extern ui_browser_window_t   ui_browser_window_qt;
-extern ui_msg_window_t       ui_msg_window_qt;
-extern ui_window_t           ui_window_qt;
-extern ui_application_t      ui_application_qt;
-
 extern ui_companion_driver_t ui_companion_cocoa;
 extern ui_companion_driver_t ui_companion_cocoatouch;
 extern ui_companion_driver_t ui_companion_qt;
